Test progress

+++ src/ported/CpuTests.h
@@ -34,7 +34,7 @@ namespace cpu {
     std::vector<std::string> LoadFenDb(const std::string &filename);
 
     void TestMoveGenPerfCPU(const std::vector<std::string> &fens, uint32_t maxDepth, uint32_t retries, uint32_t threads,
-                            const std::vector<uint64_t> &seeds);
+                            const std::vector<uint32_t> &seeds);
 } // namespace cpu
